command-center/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
exceptiongroup==1.3.1
fastapi==0.127.0
finnhub-python==2.4.26
frozendict==2.4.7
h11==0.16.0
idna==3.11
multitasking==0.0.12
numpy==2.2.6
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.1
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
websockets==15.0.1
yfinance==1.0
